ビットコイン関係のツイート取得練習完了

+++ database/migrations/2020_02_15_155314_create_tweets_table.php
@@ -15,7 +15,7 @@ class CreateTweetsTable extends Migration
     {
         Schema::create('tweets', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('tweet_id')->nullable()->comment('ツイートID');
+            $table->bigInteger('tweet_id')->nullable()->comment('ツイートID');
             $table->string('text')->nullable()->comment('ツイート内容');
             $table->dateTime('tweet_created_at')->nullable()->comment('ツイート日時');
             $table->string('screen_name')->nullable()->comment('@スクリーンネーム');
